v31.42

+++ ablogcms/php/ACMS/POST/Module/Index/Delete.php
@@ -1,5 +1,8 @@
 <?php
 
+use Acms\Services\Facades\Module;
+use Acms\Services\Facades\Logger;
+
 class ACMS_POST_Module_Index_Delete extends ACMS_POST_Module_Delete
 {
     function post()
@@ -7,13 +10,7 @@ class ACMS_POST_Module_Index_Delete extends ACMS_POST_Module_Delete
         $this->Post->reset(true);
         $this->Post->setMethod('checks', 'required');
 
-        if (enableApproval(BID, CID)) {
-            $this->Post->setMethod('module', 'operative', sessionWithApprovalAdministrator(BID, CID));
-        } elseif (roleAvailableUser()) {
-            $this->Post->setMethod('module', 'operative', roleAuthorization('module_edit', BID));
-        } else {
-            $this->Post->setMethod('module', 'operative', sessionWithContribution());
-        }
+        $this->Post->setMethod('module', 'operative', Module::canBulkDelete(BID));
         $this->Post->validate(new ACMS_Validator());
 
         $targetModules = [];
@@ -23,18 +20,26 @@ class ACMS_POST_Module_Index_Delete extends ACMS_POST_Module_Delete
             foreach ($this->Post->getArray('checks') as $mid) {
                 $id = preg_split('@:@', $mid, 2, PREG_SPLIT_NO_EMPTY);
                 $mid = $id[1];
+                if (!($mid = intval($mid))) {
+                    continue;
+                }
                 $module = loadModule($mid);
+                $moduleBlogId = (int)$module->get('blog_id');
+                if (!Module::canDelete($moduleBlogId)) {
+                    // モジュールの削除が許可されていない
+                    continue;
+                }
                 $this->delete($mid);
 
                 $targetModules[] = $module->get('label') . '（' . $module->get('identifier') . '）';
             }
             $this->Post->set('refreshed', 'refreshed');
 
-            AcmsLogger::info('選択したモジュールIDを削除しました', [
+            Logger::info('選択したモジュールIDを削除しました', [
                 'targetModules' => $targetModules,
             ]);
         } else {
-            AcmsLogger::info('選択したモジュールIDの削除に失敗しました');
+            Logger::info('選択したモジュールIDの削除に失敗しました');
         }
 
         return $this->Post;
